+++ src/main/java/pl/cleankod/ApplicationInitializer.java
@@ -1,6 +1,5 @@
 package pl.cleankod;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import feign.Feign;
 import feign.httpclient.ApacheHttpClient;
 import feign.jackson.JacksonDecoder;
@@ -13,12 +12,13 @@ import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.cache.concurrent.ConcurrentMapCacheManager;
 import org.springframework.cache.interceptor.KeyGenerator;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Primary;
+import org.springframework.context.annotation.Import;
 import org.springframework.core.env.Environment;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.scheduling.annotation.Scheduled;
 import pl.cleankod.exchange.cache.ExchangeRatesCacheKeyGenerator;
-import pl.cleankod.exchange.configuration.CustomObjectMapper;
+import pl.cleankod.exchange.configuration.CacheConfig;
+import pl.cleankod.exchange.configuration.JacksonConfig;
 import pl.cleankod.exchange.core.gateway.AccountRepository;
 import pl.cleankod.exchange.core.gateway.CurrencyConversionService;
 import pl.cleankod.exchange.core.usecase.AccountService;
@@ -36,6 +36,7 @@ import java.util.Currency;
 @EnableAutoConfiguration
 @EnableCaching
 @EnableScheduling
+@Import({JacksonConfig.class, CacheConfig.class})
 public class ApplicationInitializer {
     public static void main(String[] args) {
         SpringApplication.run(ApplicationInitializer.class, args);
@@ -91,30 +92,5 @@ public class ApplicationInitializer {
     ExceptionHandlerAdvice exceptionHandlerAdvice() {
         return new ExceptionHandlerAdvice();
     }
-
-    @Bean
-    @Primary
-    public ObjectMapper objectMapper() {
-        ObjectMapper objectMapper = new CustomObjectMapper();
-        // configure your object mapper here
-        return objectMapper;
-    }
-
-    @Bean
-    public CacheManager cacheManager() {
-        return new ConcurrentMapCacheManager("rates");
-    }
-
-    @Bean
-    public KeyGenerator exchangeRatesCacheKeyGenerator() {
-        return new ExchangeRatesCacheKeyGenerator();
-    }
-
-
-    @Scheduled(fixedRate = 60 * 60 * 1000) // every 60 minutes
-    public void evictCache() {
-        cacheManager().getCache("rates").clear();
-    }
-
 
 }
