+++ src/main/java/pl/cleankod/exchange/configuration/CustomObjectMapper.java
@@ -4,15 +4,15 @@ import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.Version;
-import com.fasterxml.jackson.databind.*;
+import com.fasterxml.jackson.databind.DeserializationContext;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializerProvider;
 import com.fasterxml.jackson.databind.deser.std.StdDeserializer;
 import com.fasterxml.jackson.databind.module.SimpleModule;
 import com.fasterxml.jackson.databind.ser.std.StdSerializer;
 import pl.cleankod.exchange.core.domain.Account;
 
 import java.io.IOException;
-import java.util.UUID;
-import java.util.regex.Pattern;
 
 public class CustomObjectMapper extends ObjectMapper {
