+++ src/main/java/pl/cleankod/exchange/core/domain/Account.java
@@ -7,7 +7,7 @@ import java.util.regex.Pattern;
 
 public record Account(Id id, Number number, Money balance) {
 
-    public static record Id(UUID value) {
+    public record Id(UUID value) {
         public Id {
             Preconditions.requireNonNull(value);
         }
@@ -22,7 +22,7 @@ public record Account(Id id, Number number, Money balance) {
         }
     }
 
-    public static record Number(String value) {
+    public record Number(String value) {
         private static final Pattern PATTERN =
                 Pattern.compile("\\d{2}[ ]?\\d{4}[ ]?\\d{4}[ ]?\\d{4}[ ]?\\d{4}[ ]?\\d{4}[ ]?\\d{4}");
