+++ src/main/java/pl/cleankod/exchange/provider/nbp/ExchangeRatesNbpClient.java
@@ -5,8 +5,10 @@ import feign.RequestLine;
 import org.springframework.cache.annotation.Cacheable;
 import pl.cleankod.exchange.provider.nbp.model.RateWrapper;
 
+import static pl.cleankod.exchange.configuration.CacheConfig.RATES_CACHE;
+
 public interface ExchangeRatesNbpClient {
-    @Cacheable(value = "rates", keyGenerator = "exchangeRatesCacheKeyGenerator")
+    @Cacheable(value = RATES_CACHE, keyGenerator = "exchangeRatesCacheKeyGenerator")
     @RequestLine("GET /exchangerates/rates/{table}/{currency}/2022-02-08")
     RateWrapper fetch(@Param("table") String table, @Param("currency") String currency);
 }
